backend/requirements.txt
beautifulsoup4
requests
psycopg2-binary
pydantic
python-dotenv
fastapi
uvicorn
PyPDF2
yfinance
pandas
torch
transformers
scikit-learn
